Update causal_read_test.sh
I have updated this script to do a few things, the first change is to make sure the table exists before running the script, and secondly, I have modified the line that inserts the data.

This is possibly not casual reads, as I have set the process to run in the background, which means it may not be finished before the read line executes.

It does highlight the issue that could happen from an application and the Script runs and executes with failures. On three runs I get 287, 331 and 291 Failures out of 1000.

But the failure is really due to the fact the insert is on a separate thread. The Script as it was was always going to complete the insert before moving on to the read. 

Doing some tuning does not help with the failure rate, three runs: 368, 393 and 384.

You can see the & at the end of line 48. This makes it a background thread - technically this should be removed - but then we never get any failures !